Move styles into new folder
