+++ package.json
@@ -41,8 +41,6 @@
   "devDependencies": {
     "@babel/preset-env": "^7.21.4",
     "@babel/preset-react": "^7.18.6",
-    "babel-jest": "^29.5.0",
-    "jest": "^29.5.0",
     "react-test-renderer": "^18.2.0"
   }
 }
