+++ src/App.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react'
-import './App.css';
+import '../src/styles/App.css';
 
 
 //components
