+++ src/styles/GraphBoard.css
