+++ src/styles/StatisticsBoard.css
