+++ src/styles/StatsTotals.css
